A mettre a jour sur Tiweb
Restant Paneau admin et images

+++ admin/admin-tribe-list.php
@@ -2,11 +2,13 @@
 
 include "../database.php";
 
+
+$msg = $_GET['x'];
+
 $SQL_REQUEST = "SELECT id_tribe, name, summary, logo, color FROM mtgTribe;";
 $tribeListRequest = $database->prepare($SQL_REQUEST);
 $tribeListRequest->execute();
 $listeTribe = $tribeListRequest->fetchAll();
-
 $tittle = "Tribes";
 require 'admin-header.php';
 ?>
@@ -33,8 +35,8 @@ require 'admin-header.php';
                                 <textarea style="border: solid grey 1px;" readonly class="resume bg-dark text-white" rows="5" cols="15" ><?=$tribe['summary']?></textarea>
                             </div>
                             <div class="card-footer longlivecenter">
-                                <a href="edit.php?vers=<?=$tribe['id_tribe']?>" class="btn btn-success">Edit</a>
-                                <a href="delete.php?vers=<?=$tribe['id_tribe']?>" class="btn btn-danger">Delete</a>
+                                <a href="edit-tribe.php?vers=<?=$tribe['id_tribe']?>" class="btn btn-success">Edit</a>
+                                <a href="delete.php?vers=<?=$tribe['id_tribe']?>" class="btn btn-danger" onclick="return confirm('Are you sure?')">Delete</a>
                                 <a href="../detailed-tribe.php?vers=<?=$tribe['id_tribe']?>" class="btn btn-primary mt-1">See in detail</a>
                             </div>
                         </div>
@@ -47,4 +49,18 @@ require 'admin-header.php';
     </div>
 <?php
 require "../footer.php";
+switch ($msg){
+    case 1:
+        echo '<script>alert("Successfully Connected")</script>';
+        break;
+    case 2:
+        echo '<script>alert("Successfully Added")</script>';
+        break;
+    case 3:
+        echo '<script>alert("Successfully Deleted")</script>';
+        break;
+    case 4:
+        echo '<script>alert("Successfully Edited")</script>';
+        break;
+}
 ?>
